requirements.txt
streamlit==1.46.0
openai==1.90.0
anthropic==0.54.0
python-dotenv==1.1.0
textblob==0.18.0
plotly==5.22.0
